+++ RELEASENOTES.md
@@ -3,6 +3,10 @@
 ### Entfernt
 - Umstellung auf mobidam-eai-commons-starter.
 
+## Sprint 18 (01.10.2024 - 22.10.2024)
+## Hinzugefügt
+- Micrometer Metriken
+
 ## Sprint 16 (20.08.2024 - 09.09.2024)
 ### Hinzugefügt
 - Schadcode-Erkennung und Mimetype-Prüfung